+++ packages/nimble-components/src/table/components/header/styles.ts
@@ -1,7 +1,6 @@
 import { css } from '@microsoft/fast-element';
 import { display } from '@microsoft/fast-foundation';
 import {
-    applicationBackgroundColor,
     controlHeight,
     standardPadding,
     tableHeaderFont,
@@ -14,9 +13,9 @@ export const styles = css`
     :host {
         height: ${controlHeight};
         align-items: center;
-        background: ${applicationBackgroundColor};
         padding: 0px calc(${standardPadding} / 2);
         font: ${tableHeaderFont};
         color: ${tableHeaderFontColor};
+        text-transform: uppercase;
     }
 `;