+++ packages/nimble-components/src/table/types.ts
@@ -24,3 +24,37 @@ export type TableFieldValue =
 export interface TableRecord {
     [key: TableFieldName]: TableFieldValue;
 }
+
+/** Represents a single row (element) in the Table's data  */
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface TableDataRecord extends TableRecord {}
+
+/**
+ * An object whose fields are defined by a particular TableColumn, which is used by the column's
+ * cellTemplate implementation.
+ */
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface TableCellRecord extends TableRecord {}
+
+export type TableStringField<FieldName extends TableFieldName> = {
+    [name in FieldName]: string | null | undefined;
+};
+
+export interface TableCellState<
+    TCellRecord extends TableCellRecord = TableCellRecord,
+    TColumnConfig = unknown
+> {
+    cellRecord: TCellRecord;
+    columnConfig: TColumnConfig;
+}
+
+export interface TableValidity {
+    readonly duplicateRecordId: boolean;
+    readonly missingRecordId: boolean;
+    readonly invalidRecordId: boolean;
+}
+
+export interface TableRowState<TData extends TableRecord = TableRecord> {
+    record: TData;
+    id: string;
+}