+++ packages/nimble-blazor/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @ni/nimble-blazor
 
-This log was last generated on Wed, 04 Jan 2023 21:51:15 GMT and should not be manually modified.
+This log was last generated on Wed, 18 Jan 2023 22:26:41 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 10.0.7
+
+Wed, 18 Jan 2023 22:26:41 GMT
+
+### Patches
+
+- Add table row ID and validity support to Blazor ([ni/nimble@a882a4c](https://github.com/ni/nimble/commit/a882a4ce55fb60b786dfde385fa97863536c45d6))
+
+## 10.0.3
+
+Sat, 14 Jan 2023 00:16:05 GMT
+
+### Patches
+
+- Adding Blazor support for TableColumnText ([ni/nimble@86ea30b](https://github.com/ni/nimble/commit/86ea30bf6d6ed01cbe449f5ec67652b151d8d2a5))
 
 ## 9.4.0